pyproject.toml
[tool.poetry]
name = "yashita"
version = "0.1.0"
description = ""
authors = ["masoud89"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
python-decouple = "^3.8"
psycopg = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
